requirements.txt
asgiref==3.4.1
Django==3.2.12
django-autoslug==1.9.8
django-bootstrap-form==3.4
django-debug-toolbar==3.2.4
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
